Fixed spamming CTRL+Z CTRL+Y causing crash + Code cleanup
