+++ Files/Files.csproj
@@ -213,7 +213,7 @@
     <Compile Include="Helpers\Convert\StorageItemHelpers.cs" />
     <Compile Include="Helpers\DialogDisplayHelper.cs" />
     <Compile Include="Helpers\DispatcherHelper.cs" />
-    <Compile Include="Helpers\EnumerableHelpers.cs" />
+    <Compile Include="Helpers\ArrayHelpers.cs" />
     <Compile Include="Helpers\ItemsDataTemplateSelector.cs" />
     <Compile Include="Helpers\JumpListManager.cs" />
     <Compile Include="Helpers\NativeDirectoryChangesHelper.cs" />
