+++ Files/Helpers/ArrayHelpers.cs
@@ -1,6 +1,6 @@
 ﻿namespace Files.Helpers
 {
-    public class EnumerableHelpers
+    public class ArrayHelpers
     {
         public static int FitBounds(int index, int length) =>
             index >= length ? length - 1 : (index < 0 ? 0 : index);
